Made /glass convert all block from /colorcode
