+++ src/main/java/com/domain/redstonetools/features/commands/ColorCodeFeature.java
@@ -2,6 +2,7 @@ package com.domain.redstonetools.features.commands;
 
 import com.domain.redstonetools.features.Feature;
 import com.domain.redstonetools.features.arguments.Argument;
+import com.domain.redstonetools.utils.BlockColorUtils;
 import com.mojang.brigadier.Command;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import com.sk89q.worldedit.EditSession;
@@ -29,9 +30,6 @@ import static com.mojang.brigadier.arguments.BoolArgumentType.bool;
 
 @Feature(name = "Color Code", description = "Color codes all colorable blocks in your WorldEdit selection.", command = "/colorcode")
 public class ColorCodeFeature extends CommandFeature {
-    private static final java.util.regex.Pattern MATCH_TARGET_PATH_PATTERN = java.util.regex.Pattern.compile(
-    "(_wool$)|(_concrete$)|(_stained_glass$)|(_terracotta$)|(_concrete_powder$)|(_glazed_terracotta$)"
-    );
 
     public static final Argument<String> color = Argument
             .ofType(blockColor());
@@ -39,55 +37,11 @@ public class ColorCodeFeature extends CommandFeature {
             .ofType(bool())
             .withDefault(false);
 
-    private boolean shouldBeColored(World world, BlockVector3 pos, boolean onlyWhite) {
-        BlockState state = world.getBlock(pos);
-        if (state == null)
-            return false;
-
-        var blockColor = getBlockColor(state);
-        if (blockColor == null)
-            return false;
-
-        if (onlyWhite && !blockColor.equals("white"))  // TODO: Creating an enum for colors would be less error prone
-            return false;
-
-        return true;
-    }
-
-    private int getColorEndIndex(String blockId) {
-        var matcher = MATCH_TARGET_PATH_PATTERN.matcher(blockId);
-        if (!matcher.find())
-            return -1;
-
-        return matcher.start();
-    }
-
-    private String getBlockColor(BlockState state) {
-        var blockId = state.getBlockType().getId();
-
-        var colorEndIndex = getColorEndIndex(blockId);
-
-        if (colorEndIndex == -1)
-            return null;
-
-        return blockId.substring(0, colorEndIndex);
-    }
-
-    private String getColorlessBlockId(BlockState state) {
-        var blockId = state.getBlockType().getId();
-
-        var colorEndIndex = getColorEndIndex(blockId);
-
-        if (colorEndIndex == -1)
-            return null;
-
-        return blockId.substring(colorEndIndex);
-    }
 
     private BaseBlock setBlockColor(World world, BlockVector3 pos, String color) {
         BlockState state = world.getBlock(pos);
 
-        var colorlessBlockId = getColorlessBlockId(state);
+        var colorlessBlockId = BlockColorUtils.getColorlessBlockId(state);
         var coloredBlockId = "minecraft:" + color + colorlessBlockId;
 
         BlockType blockType = BlockType.REGISTRY.get(coloredBlockId);
@@ -126,7 +80,7 @@ public class ColorCodeFeature extends CommandFeature {
                 new Mask() {
                     @Override
                     public boolean test(BlockVector3 vector) {
-                        return shouldBeColored(world, vector, onlyWhite.getValue());
+                        return BlockColorUtils.shouldBeColored(world, vector, onlyWhite.getValue());
                     }
 
                     @Nullable
